+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 - The returned dataframe has its columns properly cast to the detected types [#104](https://github.com/datagouv/csv-detective/pull/104)
 - Raise an error if the encoding could not be guessed [#106](https://github.com/datagouv/csv-detective/pull/106)
 - Fix CLI and minio routine [#107](https://github.com/datagouv/csv-detective/pull/107)
+- Allow to only specify tests to skip ("all but...") [#108](https://github.com/datagouv/csv-detective/pull/108)
+- Fix bool casting [#109](https://github.com/datagouv/csv-detective/pull/109)
+- Handle csv.gz files [#110](https://github.com/datagouv/csv-detective/pull/110)
+- Refactor file tests [#110](https://github.com/datagouv/csv-detective/pull/110)
+- Restructure repo (breaking changes) [#111](https://github.com/datagouv/csv-detective/pull/111)
+- Better float detection [#113](https://github.com/datagouv/csv-detective/pull/113)
 
 ## 0.7.4 (2024-11-15)